auth: show _OR_ during login only when other methods are available.
There might be good reasons to have other external authentication
methods such as SAML configured, but none of them is available. This
happens for example when you already enable SAML so that Zulip is able
to generate the metadata in XML format, but you haven't configured an
IdP yet. This commit makes sure that the phrase _OR_ is only shown on
the login/account page when there are actually other authentication
methods available. When they are just configured, but not available
yet, the page looks like as if no other authentication methods would be
configured.