Load django from /home/zulip/deployments/current
This commit is based on a brief discussion with Tim Abbott. In a
nutshell, the current version failed to load django when the
process_fts_updates script was outside the development root and would
then not be able to contact the Postgresql server properly. Thsi commit
fixes that by trying to load django from
/home/zulip/deployments/current and potentially the venv in
zulip-py3-venv as well.

This might not be the optimal solution here, but the optimal solution
may depend on what kind of environments the process_fts_updates might
be executed in.